Rename set_1 module to windows for consistency
